Analyse

+++ src/Session/Session.java
@@ -27,7 +27,7 @@ public class Session {
 		this.manager = manager;
 		debutTime = new Date(System.currentTimeMillis());
 		currentTime = new Date(System.currentTimeMillis());
-		shceduledEnd = new Date (debutTime.getTime()+(long)(60*1000)); // fin de la session prévue 10 min après le début
+		shceduledEnd = new Date (debutTime.getTime()+60*1000); // fin de la session prévue 10 min après le début
 		running = true;
 		timeUpdater = new TimeUpdater(this);
 		passwordTries = new ArrayList<PasswordTry>();
@@ -64,7 +64,7 @@ public class Session {
 	}
 	
 	public void reshceduleEnd(){
-		shceduledEnd = new Date(shceduledEnd.getTime() + (long)(20*1000));
+		shceduledEnd = new Date(shceduledEnd.getTime() + 20*1000);
 		System.out.println("End of session reshceduled");
 	}
